some style refinements
