+++ packages/ui/components/pages/Layout/FusePageLayout.tsx
@@ -1,7 +1,7 @@
 import { Flex } from '@chakra-ui/react';
 
 import Footer from '@ui/components/pages/Layout/Footer';
-import { FuseNavbar } from '@ui/components/pages/Layout/FuseNavbar';
+import { MidasNavbar } from '@ui/components/pages/Layout/MidasNavbar';
 import { useColors } from '@ui/hooks/useColors';
 import { FusePageLayoutProps } from '@ui/types/ComponentPropsType';
 
@@ -16,7 +16,7 @@ const FusePageLayout = ({ children }: FusePageLayoutProps) => {
       bgColor={cPage.primary.bgColor}
       justifyContent="flex-start"
     >
-      <FuseNavbar />
+      <MidasNavbar />
       {children}
       <Footer />
     </Flex>
